@@ -43,7 +43,7 @@ export default Vue.extend({
     offset-y
     min-width="290px"
   >
-    <template v-slot:activator="{ on, attrs }">
+    <template #activator="{ on, attrs }">
       <v-text-field
         :value="formattedDate"
         label="Choose"
